--- src/main/java/Command.java
@@ -1,11 +1,0 @@
-public abstract class Command {
-    protected boolean isDone = false;
-
-    public void execute(TaskList tasklist, Ui ui, Storage store) {
-
-    }
-
-    public boolean isExit() {
-        return isDone;
-    }
-}
