+++ src/main/java/duke/Ui.java
@@ -1,4 +1,6 @@
+package duke;
 import java.util.Scanner;
+import duke.task.*;
 
 public class Ui {
     private final Scanner sc;
