+++ src/main/java/duke/command/DoneCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class DoneCommand extends Command {
     public DoneCommand() {
         this.isDone = true;
