+++ src/main/java/duke/task/Deadline.java
@@ -1,23 +1,24 @@
+package duke.task;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Deadline extends Task{
     private final LocalDate date;
 
-    public Event(String label, LocalDate date) {
+    public Deadline(String label, LocalDate date) {
         this.date = date;
         this.label = label;
     }
 
     @Override
     public String getType() {
-        return "E";
+        return "D";
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() +
-                "(at: " + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
+        return "[D]" + super.toString() +
+                "(by: " + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 
     @Override
