+++ src/main/java/duke/task/Event.java
@@ -1,23 +1,24 @@
+package duke.task;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task{
+public class Event extends Task{
     private final LocalDate date;
 
-    public Deadline(String label, LocalDate date) {
+    public Event(String label, LocalDate date) {
         this.date = date;
         this.label = label;
     }
 
     @Override
     public String getType() {
-        return "D";
+        return "E";
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() +
-                "(by: " + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
+        return "[E]" + super.toString() +
+                "(at: " + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 
     @Override
